feat: include projects directory and project detail pages in private
- Added a new Projects Directory page for marketers, featuring a searchable and filterable grid of projects.
- Created a reusable Project Detail component to display detailed project information, including stats and application options.
- Updated existing project pages to utilize the new ProjectDetail component, enhancing consistency across public and private contexts.
- Introduced navigation links and improved layout for better user experience in the marketer dashboard.
